# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  handlers.logoutRequest:
    properties:
      refresh_token:
        type: string
    type: object
...
  handlers.refreshRequest:
    properties:
      refresh_token:
        type: string
    type: object
...
  models.AuthResponse:
    properties:
      profile:
        $ref: '#/definitions/models.Profile'
      refresh_token:
        type: string
      token:
        type: string
      user:
        $ref: '#/definitions/models.UserResponse'
    type: object
...
  models.LoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    type: object
...
  models.Profile:
    properties:
      avatar:
        type: string
      bio:
        type: string
      cover_image:
        type: string
      created_at:
        type: string
      id:
        type: string
      name:
        type: string
      role:
        type: string
      settings:
        $ref: '#/definitions/models.ProfileSettings'
      social_links:
        $ref: '#/definitions/models.SocialLinks'
      updated_at:
        type: string
      user_id:
        type: string
    type: object
  models.ProfileSettings:
    properties:
      currency:
        type: string
      date_format:
        description: '"DD/MM/YYYY", "MM/DD/YYYY", etc'
        type: string
      first_day_of_week:
        description: 0=Sunday, 1=Monday
        type: integer
      language:
        type: string
      theme:
        $ref: '#/definitions/models.ThemeSettings'
    type: object
  models.RegisterRequest:
    properties:
      email:
        type: string
      name:
        type: string
      password:
        type: string
    type: object
...
  models.SocialLinks:
    properties:
      github:
        type: string
      instagram:
        type: string
      linkedin:
        type: string
      twitter:
        type: string
      website:
        type: string
    type: object
...
  models.ThemeSettings:
    properties:
      accent_color:
        description: Hex color for accents
        type: string
      mode:
        description: '"dark", "light", "system"'
        type: string
      primary_color:
        description: 'Hex color, default: #D4AF37 (gold)'
        type: string
    type: object
...
  models.UserResponse:
    properties:
      created_at:
        type: string
      email:
        type: string
      id:
        type: string
    type: object
host: localhost:8088
info:
  contact: {}
  description: API de autenticacao e gerenciamento de usuarios
  title: Tron Legacy API
  version: "1.0"
paths:
...
  /auth/login:
    post:
      consumes:
      - application/json
      description: Autentica usuário com email e senha
      parameters:
      - description: Dados de login
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.AuthResponse'
        "400":
          description: Invalid request body
          schema:
            type: string
        "401":
          description: Invalid credentials
          schema:
            type: string
      summary: Login de usuário
      tags:
      - auth
  /auth/logout:
    post:
      consumes:
      - application/json
      description: Invalida o refresh token fornecido
      parameters:
      - description: Refresh token para invalidar
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.logoutRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Logout do usuário
      tags:
      - auth
  /auth/me:
    get:
      consumes:
      - application/json
      description: Retorna informações do usuário autenticado
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.AuthResponse'
        "401":
          description: Unauthorized
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Dados do usuário logado
      tags:
      - auth
...
  /auth/refresh:
    post:
      consumes:
      - application/json
      description: Troca um refresh token válido por um novo par de access + refresh
        token
      parameters:
      - description: Refresh token
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.refreshRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.AuthResponse'
        "400":
          description: refresh_token is required
          schema:
            type: string
        "401":
          description: Invalid or expired refresh token
          schema:
            type: string
      summary: Renovar tokens de acesso
      tags:
      - auth
  /auth/register:
    post:
      consumes:
      - application/json
      description: Cria uma nova conta com email e senha
      parameters:
      - description: Dados de registro
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.AuthResponse'
        "400":
          description: Invalid request body
          schema:
            type: string
        "409":
          description: Email already exists
          schema:
            type: string
      summary: Registrar novo usuário
      tags:
      - auth
...
securityDefinitions:
  BearerAuth:
    description: 'Digite: Bearer {seu_token_aqui}'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"